grammar: create new branch for changing whitespace rules.
It appears that we can't let the parser ignore space and tab characters. constructs like 'do n=1,2' can't be processed because the 'do n' gets compressed into 'don' and the parser considers this an assignment statement. Prevent the parser from skipping whitespace, and explicitly handle space and tab characters. This is the first pass with a couple of passing tests, all other existing tests are expected fails.