testing: major refactor of the testing.
Instead of having many F03 test files, they have been removed and the source code placed into strings in the python test functions. The code snippet is written to a temporary file. If the code snippet is compilable (e.g. a module or subroutine) it is run through gfortran in pedantic f2003 syntax checking mode. The code is then run through the grako parser. The AST is checked, and written to a file. The new and original files are compared to unsure we can recreate the original input.